backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.2
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5
channels>=4.0
channels-redis>=4.2
django-redis>=5.4
argon2-cffi>=23.1
django-storages>=1.14
boto3>=1.34
whitenoise>=6.6
gunicorn>=21.2
pandas>=2.2
numpy>=1.26
